:recycle: Move auto_width and auto_height to their own textlayout module
