+++ render-wasm/src/wasm/text.rs
@@ -1,8 +1,7 @@
 use crate::mem;
-use crate::shapes::{auto_height, build_paragraphs_with_width, GrowType, RawTextData, Type};
-
-use crate::STATE;
-use crate::{with_current_shape, with_current_shape_mut};
+use crate::shapes::{GrowType, RawTextData, Type};
+use crate::textlayout::{auto_height, build_paragraphs_with_width};
+use crate::{with_current_shape, with_current_shape_mut, STATE};
 
 #[no_mangle]
 pub extern "C" fn clear_shape_text() {
